requirements.txt
# Install
wheel

# Run
chromatictools
numpy
scipy
scikit-learn

# Test
cython
more-itertools

# Notebooks
jupyter
matplotlib
librosa
requests

# GUI
ttkthemes
Pillow
pygame
throttle

# Style
pylint
yapf

# Coverage
coverage

# Docs
sphinx
sphinx_rtd_theme
m2r2
recommonmark

# Packaging
twine
pyinstaller
